change code dependency direction
